refactors as suggested by PR feedback
