+++ src/errors/mod.rs
@@ -1,5 +1,3 @@
-use core::fmt;
-
 use pyo3::prelude::*;
 
 mod line_error;
@@ -31,53 +29,4 @@ pub fn py_err_string(py: Python, err: PyErr) -> String {
         },
         Err(_) => "Unknown Error".to_string(),
     }
-}
-
-// TODO: is_utf8_char_boundary, floor_char_boundary and ceil_char_boundary
-// with builtin methods once https://github.com/rust-lang/rust/issues/93743 is resolved
-// These are just copy pasted from the current implementation
-const fn is_utf8_char_boundary(value: u8) -> bool {
-    // This is bit magic equivalent to: b < 128 || b >= 192
-    (value as i8) >= -0x40
-}
-
-pub fn floor_char_boundary(value: &str, index: usize) -> usize {
-    if index >= value.len() {
-        value.len()
-    } else {
-        let lower_bound = index.saturating_sub(3);
-        let new_index = value.as_bytes()[lower_bound..=index]
-            .iter()
-            .rposition(|b| is_utf8_char_boundary(*b));
-
-        // SAFETY: we know that the character boundary will be within four bytes
-        unsafe { lower_bound + new_index.unwrap_unchecked() }
-    }
-}
-
-pub fn ceil_char_boundary(value: &str, index: usize) -> usize {
-    let upper_bound = Ord::min(index + 4, value.len());
-    value.as_bytes()[index..upper_bound]
-        .iter()
-        .position(|b| is_utf8_char_boundary(*b))
-        .map_or(upper_bound, |pos| pos + index)
-}
-
-pub fn write_truncated_to_limited_bytes<F: fmt::Write>(f: &mut F, val: &str, max_len: usize) -> std::fmt::Result {
-    if val.len() > max_len {
-        let mid_point = (max_len as f64 / 2.0).ceil() as usize;
-        write!(
-            f,
-            "{}...{}",
-            &val[0..floor_char_boundary(val, mid_point)],
-            &val[ceil_char_boundary(val, val.len() - (mid_point - 1))..]
-        )
-    } else {
-        write!(f, "{val}")
-    }
-}
-
-// preserved for backwards compatibility, can be removed in a major (or potentially minor) release
-pub fn write_truncated_to_50_bytes<F: fmt::Write>(f: &mut F, val: &str) -> std::fmt::Result {
-    write_truncated_to_limited_bytes(f, val, 50)
 }
